requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pillow
ultralytics
torch
torchvision
transformers
qdrant-client[fastembed]
pymupdf